Initial commit of the Pastry simulation project.
In this commit, we upload the original files without any modification.
It's kept as what it was like in late 2009, when I was still an M.Phil
candidate in HKU CS department. The code style at that time may not
good, which can be improved by future commits if time permits.